split interaction endpoint to new router, updated calendar.
