+++ lib/app.js
@@ -7,6 +7,7 @@ app.use(express.urlencoded({extended:false}));
 // app.use('/slack/zoom', require('./controllers/zoom'));
 // app.use('/slack/canvas', require('./controllers/canvas'));
 app.use('/slack/calendar', require('./controllers/calendar'));
+app.use('/slack/interaction', require('./controllers/interaction'));
 // app.use('/slack/npm', require('./controllers/npm'));
 // app.use('/slack/futurama', require('./controllers/fun-stuff'));
